requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

accelerate==1.12.0
aiofiles==24.1.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
av==16.1.0
beautifulsoup4==4.14.3
boto3==1.42.28
botocore==1.42.28
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dill==0.4.0
distro==1.9.0
doclayout_yolo==0.0.4
Elixir==0.7.1
fast-langdetect==0.2.5
fastapi==0.128.0
fasttext-predict==0.9.2.4
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
ftfy==6.3.1
gradio==5.49.1
gradio_client==1.13.3
gradio_pdf==0.0.22
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-retries==0.4.5
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
json_repair==0.55.0
kiwisolver==1.4.9
lazy_loader==0.4
loguru==0.7.3
magika==1.0.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mineru==2.7.1
mineru_vl_utils==0.1.21
modelscope==1.33.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.20.1
openai==2.15.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pbr==6.1.0
pdfminer.six==20260107
pdftext==0.6.3
pillow==11.3.0
polars==1.37.1
polars-runtime-32==1.37.1
protobuf==6.33.4
psutil==7.2.1
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.7
pyparsing==3.3.1
pypdf==6.6.0
pypdfium2==4.30.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
rich==14.2.0
robust-downloader==0.0.2
ruff==0.14.12
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sqlalchemy-migrate==0.11.0
sqlparse==0.5.2
starlette==0.50.0
stringzilla==4.6.0
sympy==1.13.1
Tempita==0.6.0
thop==0.1.1.post2209072238
tifffile==2026.1.14
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.57.5
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.2
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.2.14
websockets==15.0.1
win32_setctime==1.2.0
xlwt==0.7.2
